http_responses: add 'app code' in error messages
Allow CGI scripts to set custom numeric application-code which
will appear before the random request code, example:
     set_app_code(555)
then, when an error occurs:
     request 555.0123456789: something went wrong

The 555 identifies the application (i.e. the CGI script),
the other digits are random number which will be shown to the user
when an error occurs, but also written to the log (possibly with more
detailed error information).